handled the deduping off cacheable requests via WaitingRequestManager WaitingRequestManager instance is created in RequestQueue setup and passed to CacheDispatcher threads
